+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 - Potential fix for setting partition clearing the slot in another row, in some cases, due to a race condition in cell state between clicks (usually when world was just loaded and the cell state was not fully synced yet).
 - Fix upgrades not being clickable directly on cells.
 - Fix (virtual) slots not being interactible with JEI.
+- Fix storage bus contents being truncated to the partition slot count
 
 
 ## [1.5.3] - 2026-03-25